Final changes for the competition

+++ KaggleCompVinBD1/src/utils/collater.py
@@ -29,7 +29,7 @@ class SimpleCollater(Collater):
         elif 'annotations' in batch[0]:
             annotations = [{
                 'boxes': torch.tensor(x['annotations']['boxes']).to(config.devices[0]),
-                'labels': torch.tensor(x['annotations']['labels'], dtype=torch.float).to(config.devices[0]),
+                'labels': torch.tensor(x['annotations']['labels'], dtype=torch.long).to(config.devices[0]),
                 'category_ids': [] if 'category_ids' not in x['annotations'] else torch.tensor(x['annotations']['category_ids']).to(config.devices[0]),
             } for x in batch]
